Finalize interface_stage_2
