+++ front_panel/public/index.html
@@ -54,6 +54,5 @@
             </div>
         </div>
         <script type="module" src="/main.js"></script>
-        <script type="module" src="/control2.js"></script>
     </body>
 </html>
