+++ front_panel/src/control2/button.js
@@ -1,8 +1,21 @@
+import { getDisplay2 } from "../dsiplay2/display2";
+
 export class Button3D {
-    constructor(model, action, mixer) {
+    constructor(number, model, action, mixer, label) {
+        this.number = number;
         this.model = model;
         this.action = action;
         this.mixer = mixer;
+        this.label = label;
+        this.doing = null;
+    }
+
+    addDoing(callback) {
+        if (callback && typeof callback === "function") {
+            this.doing = callback;
+        } else {
+            console.warn("Attempt to assing not a function to button doing.");
+        }
     }
 
     press() {
@@ -11,14 +24,21 @@ export class Button3D {
             this.action.play();
             this.action.reset();
         }
+        if (this.doing) {
+            this.doing();
+        } else {
+            const display2 = getDisplay2();
+            display2.write_line("No function assigned to this button");
+        }
     }
 }
 
 export let buttons = [];
 
-export function addButton3D(model, action, mixer) {
-    const button3D = new Button3D(model, action, mixer);
+export function addButton3D(number, model, action, mixer, label) {
+    const button3D = new Button3D(number, model, action, mixer, label);
     buttons.push(button3D);
+    buttons.sort((a, b) => b.number - a.number);
 }
 
 export function updateButton3DMixers(delta) {
