+++ front_panel/src/control2/toggle_switch.js
@@ -1,17 +1,23 @@
 import { animateEmission } from "./indicator";
+import { getDisplay2 } from "../dsiplay2/display2";
 
 export class ToggleSwitch {
-    constructor(model, action, mixer, state) {
+    constructor(number, model, action, mixer, state, label) {
+        this.number = number;
         this.model = model;
         this.action = action;
         this.mixer = mixer;
         this.state = state;
+        this.label = label;
         this.doing = null;
     }
 
     addDoing(callback) {
-        console.log("Callback: " + callback);
-        this.doing = callback;
+        if (callback && typeof callback === "function") {
+            this.doing = callback;
+        } else {
+            console.warn("Attempt to assing not a function to t_switch doing.");
+        }
     }
 
     toggle() {
@@ -21,15 +27,28 @@ export class ToggleSwitch {
             this.action.paused = false;
             this.action.play();
         }
-        this.doing();
+        if (this.doing) {
+            this.doing();
+        } else {
+            const display2 = getDisplay2();
+            display2.write_line("No function assigned to this t_switch");
+        }
     }
 }
 
 export let toggle_switches = [];
 
-export function addToggleSwitch(model, action, mixer) {
-    const toggleSwitch = new ToggleSwitch(model, action, mixer, false);
+export function addToggleSwitch(i, model, action, mixer, label) {
+    const toggleSwitch = new ToggleSwitch(
+        i,
+        model,
+        action,
+        mixer,
+        false,
+        label
+    );
     toggle_switches.push(toggleSwitch);
+    toggle_switches.sort((a, b) => a.number - b.number);
 }
 
 export function updateToggleMixers(delta) {
