+++ front_panel/webpack.config.js
@@ -3,10 +3,9 @@ const path = require("path");
 module.exports = {
     entry: {
         main: "./src/script.js",
-        control2: "./src/control2/main_control2.js",
     },
     output: {
-        filename: "[name].js",
+        filename: "main.js",
         path: path.resolve(__dirname, "dist"),
     },
 };
